refactor: FcmTokens 생성 로직 Event로 추출
