+++ backend/src/main/java/com/zzang/chongdae/event/domain/SaveCommentEvent.java
@@ -1,6 +1,8 @@
 package com.zzang.chongdae.event.domain;
 
 import com.zzang.chongdae.comment.repository.entity.CommentEntity;
+import com.zzang.chongdae.member.repository.entity.MemberEntity;
+import com.zzang.chongdae.notification.domain.FcmTokens;
 import com.zzang.chongdae.offeringmember.repository.entity.OfferingMemberEntity;
 import java.util.List;
 import lombok.Getter;
@@ -10,11 +12,19 @@ import org.springframework.context.ApplicationEvent;
 public class SaveCommentEvent extends ApplicationEvent {
 
     private final CommentEntity comment;
-    private final List<OfferingMemberEntity> offeringMembers;
+    private final FcmTokens tokens;
 
     public SaveCommentEvent(Object source, CommentEntity comment, List<OfferingMemberEntity> offeringMembers) {
         super(source);
         this.comment = comment;
-        this.offeringMembers = offeringMembers;
+        this.tokens = createFcmTokens(comment.getMember(), offeringMembers);
+    }
+
+    private FcmTokens createFcmTokens(MemberEntity writer, List<OfferingMemberEntity> offeringMembers) {
+        List<MemberEntity> membersNotWriter = offeringMembers.stream()
+                .map(OfferingMemberEntity::getMember)
+                .filter(member -> !member.isSame(writer))
+                .toList();
+        return FcmTokens.from(membersNotWriter);
     }
 }
