+++ backend/src/main/java/com/zzang/chongdae/notification/service/FcmNotificationService.java
@@ -6,6 +6,7 @@ import com.google.firebase.messaging.MulticastMessage;
 import com.zzang.chongdae.comment.repository.entity.CommentEntity;
 import com.zzang.chongdae.member.repository.entity.MemberEntity;
 import com.zzang.chongdae.notification.domain.FcmToken;
+import com.zzang.chongdae.notification.domain.FcmTokens;
 import com.zzang.chongdae.notification.domain.FcmTopic;
 import com.zzang.chongdae.notification.service.message.CommentMessageManager;
 import com.zzang.chongdae.notification.service.message.OfferingMessageManager;
@@ -13,7 +14,6 @@ import com.zzang.chongdae.notification.service.message.ParticipationMessageManag
 import com.zzang.chongdae.notification.service.message.RoomStatusMessageManager;
 import com.zzang.chongdae.offering.repository.entity.OfferingEntity;
 import com.zzang.chongdae.offeringmember.repository.entity.OfferingMemberEntity;
-import java.util.List;
 import javax.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -61,11 +61,11 @@ public class FcmNotificationService {
     }
 
     @Nullable
-    public BatchResponse saveComment(CommentEntity comment, List<OfferingMemberEntity> offeringMembers) {
-        MulticastMessage message = commentMessageManager.messageWhenSaveComment(comment, offeringMembers);
-        if (message == null) {
+    public BatchResponse saveComment(CommentEntity comment, FcmTokens tokens) {
+        if (tokens.isEmpty()) {
             return null;
         }
+        MulticastMessage message = commentMessageManager.messageWhenSaveComment(comment, tokens);
         return notificationSender.send(message);
     }
