+++ backend/src/main/java/com/zzang/chongdae/notification/service/message/CommentMessageManager.java
@@ -2,12 +2,8 @@ package com.zzang.chongdae.notification.service.message;
 
 import com.google.firebase.messaging.MulticastMessage;
 import com.zzang.chongdae.comment.repository.entity.CommentEntity;
-import com.zzang.chongdae.member.repository.entity.MemberEntity;
 import com.zzang.chongdae.notification.domain.FcmData;
 import com.zzang.chongdae.notification.domain.FcmTokens;
-import com.zzang.chongdae.offeringmember.repository.entity.OfferingMemberEntity;
-import java.util.List;
-import javax.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -20,25 +16,12 @@ public class CommentMessageManager {
 
     private final FcmMessageCreator messageCreator;
 
-    @Nullable
-    public MulticastMessage messageWhenSaveComment(CommentEntity comment, List<OfferingMemberEntity> offeringMembers) {
-        FcmTokens tokens = createFcmTokens(comment.getMember(), offeringMembers);
-        if (tokens.isEmpty()) {
-            return null;
-        }
+    public MulticastMessage messageWhenSaveComment(CommentEntity comment, FcmTokens tokens) {
         FcmData data = new FcmData();
         data.addData("title", comment.getOffering().getTitle());
         data.addData("body", MESSAGE_BODY_FORMAT.formatted(comment.getMember().getNickname(), comment.getContent()));
         data.addData("offering_id", comment.getOffering().getId());
         data.addData("type", MESSAGE_TYPE);
         return messageCreator.createMessages(tokens, data);
-    }
-
-    private FcmTokens createFcmTokens(MemberEntity writer, List<OfferingMemberEntity> offeringMembers) {
-        List<MemberEntity> membersNotWriter = offeringMembers.stream()
-                .map(OfferingMemberEntity::getMember)
-                .filter(member -> !member.isSame(writer))
-                .toList();
-        return FcmTokens.from(membersNotWriter);
     }
 }
